refactor new tests directory structure
