+++ src/components/test-new/locations-input/LocationsInput.jsx
